devlopment/segmentation/fastsam-nn/requirements.txt
absl-py==2.0.0
aiofiles==0.7.0
asgiref==3.4.1
asttokens==2.4.1
backcall==0.2.0
cachetools==5.3.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
comm==0.1.4
cycler==0.10.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.68.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==6.8.0
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
jupyterlab-widgets==3.0.9
kiwisolver==1.3.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.4.3
matplotlib-inline==0.1.6
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==3.4.18.65
pandas==1.3.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5  
Pillow>=10.0.1
setuptools>=65.5.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.8.2
Pygments==2.16.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
starlette==0.14.2
tensorboard==2.15.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.0
tqdm==4.66.1
traitlets==5.12.0
typing-extensions==3.10.0.2
ultralytics==8.0.202
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.8
websockets==10.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
